partly added room transition
partially implemented room transitions. TODO add padding to zmapper/zmap, resolve OOB issues, resolve zmap-level kinematics mismatch

+++ zdata.hpp
@@ -8,8 +8,8 @@
 
 struct zdata {
 	float z, x_slope, y_slope;
-	int room_id; //room_id is the alpha bit and is transparent from the background. 0 is reserved for inaccessible
-	zdata(std::array<float, 3> data, int room_id) :z(data[0]), x_slope(data[1]), y_slope(data[2]), room_id(room_id) {}
+	uint8_t room_id; //room_id is the alpha bit and is transparent from the background. 0 is reserved for inaccessible
+	zdata(std::array<float, 3> data, uint8_t room_id) :z(data[0]), x_slope(data[1]), y_slope(data[2]), room_id(room_id) {}
 	static zdata floor() {
 		return zdata({ -INFINITY,0,0 }, 0);
 	}
